scripts/requirements.txt
python-dotenv==1.0.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
